deprecate HeadmateListener
