+++ src/main/java/dev/tizu/headmate/ThisPlugin.java
@@ -4,7 +4,6 @@ import org.bukkit.Bukkit;
 import org.bukkit.plugin.java.JavaPlugin;
 
 import dev.tizu.headmate.editor.EditorListener;
-import dev.tizu.headmate.headmate.HeadmateListener;
 import dev.tizu.headmate.menu.Menu;
 import dev.tizu.headmate.wand.WandListener;
 
@@ -15,7 +14,6 @@ public class ThisPlugin extends JavaPlugin {
     public void onEnable() {
         instance = this;
         Bukkit.getPluginManager().registerEvents(new EditorListener(), this);
-        Bukkit.getPluginManager().registerEvents(new HeadmateListener(), this);
         Bukkit.getPluginManager().registerEvents(new Menu.MenuListener(), this);
         Bukkit.getPluginManager().registerEvents(new WandListener(), this);
     }
